### 2025-12-23 #9685 @editorialbot
[REVIEW]: JAXtronomy: A JAX port of lenstronomy
**Submitting author:** @ahuang314 (<a href="http://orcid.org/0009-0001-8629-8826">Alan Huang</a>)
**Repository:** https://github.com/lenstronomy/JAXtronomy
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.1.3
