### 2026-01-12 #9685 @ahuang314
Hi @ConnorStoneAstro, I am working on a new feature, which is currently in ahuang314/JAXtronomy but has not yet been merged into lenstronomy/JAXtronomy. The plan will be to merge this feature into lenstronomy/JAXtronomy, hopefully this week.
