### 2026-01-12 lenstronomy/JAXtronomy#70 @ConnorStoneAstro
[JOSS] Missing dependencies
Installing JAXtronomy went smoothly by simply running `pip install jaxtronomy` as described in the installation guide. However, when running some tests (specifically the unit tests in the repo) I encountered a number of missing dependencies: `cobaya`, `zeus-mcmc`, `dynesty`, `nautilus-sampler`, and `emcee`. These are listed in the `test_requirements.txt` file, but are worth mentioning in the installation guide. Perhaps a `pip install jaxtronomy[sampling]` or similar option could be set up for users? The error messages make it clear what packages are missing (except `zeus` vs `zeus-mcmc` and `nautilus` vs `nautilus-sampler`), but some require specific versions as the latest version isn't compatible. So it seems worthwhile to provide a simple solution to users.

Also, I am not sure why, but `optax` and `numpyro` didn't install on my system. I used a fresh virtual environment for Python 3.12.3, you can see the full set of commands:

```bash
~$ python3 -m venv venvs/JAXtronomy
~$ source venvs/JAXtronomy/bin/activate
(JAXtronomy) ~$ pip install jaxtronomy
(JAXtronomy) ~$ python
Python 3.12.3 (main, Nov  6 2025, 13:44:16) [GCC 13.3.0] on linux
>>> import jaxtronomy
>>> jaxtronomy.__version__
'0.1.1'
```

I was able to manually install them with pip, and the corresponding tests worked normally afterwards.

It looks like `optax` and `numpyro` are only used in the `Samplers` (and corresponding `Workflow`) submodule, so perhaps they belong with the other samplers?
