### 2026-01-12 lenstronomy/JAXtronomy#71 @ConnorStoneAstro
[JOSS] Metadata on PyPI
Looking up `jaxtronomy` on PyPI, I see that some metadata is missing: https://pypi.org/project/jaxtronomy/#description
The project description (which can just be the README for the repo) is not there, including project links such as to the GitHub, documentation, and where to submit issues. Filling these out should be fairly simple and ensures users can find the appropriate information for the package easily.
