### 2026-01-12 lenstronomy/JAXtronomy#73 @ConnorStoneAstro
[JOSS] Poor tolerance setting in example notebook
In the line below, the example L-BFGS optimizer tolerance is set to 5200:

https://github.com/lenstronomy/JAXtronomy/blob/eeb51d6beef58c975f7cb6fbf9385f6a8497c6ac/notebooks/modeling_a_simple_Einstein_ring.ipynb#L830

Assuming the noise is roughly Gaussian, this corresponds to a chi2/ndf of ~1.4 which is not low enough to consider optimization converged. At least not in the example case where our model is guaranteed to have coverage of the ground truth. Further, this limit is very specific to the example and would be invalid for any other size of image. So it could be misleading to users. In the optax documentation their example uses a tolerance on the gradient of the parameters, since the gradient is expected to be zero at the minima: https://optax.readthedocs.io/en/stable/_collections/examples/lbfgs.html#l-bfgs-solver .  It would be better to implement a more general tolerance for the solver based on gradient information, or other convergence criteria, though a simple fix would just be to add some discussion in the notebook. Either way, it would probably be best to tighten the tolerance criteria since some slight structure is still visible in the residuals.

Also, I am not sure if this is intentional for the sake of demonstration, but the PSO optimizer also does not seem to converge as there are substantial residuals remaining.
