### 2026-01-12 #9685 @ConnorStoneAstro
@ivastar , in the checklist there is a "reproducibility" check under "general checks" which asks if we can reproduce any results in the paper. I notice that the table of relative runtimes is in the paper, but this would be quite challenging to reproduce. Further, the results all seem very reasonable, and I was able to run a bunch of other examples after installing the code. Those other examples, such as this one: https://github.com/lenstronomy/JAXtronomy/blob/main/notebooks/modeling_a_simple_Einstein_ring.ipynb did include a head-to-head comparison that showed impressive speedup. So I think I am convinced. Is this enough for the review checklist?
