### 2026-01-12 lenstronomy/JAXtronomy#73 @ahuang314
Thanks for raising this issue @ConnorStoneAstro !

Indeed, a tolerance of 5200 is quite high, and the reason for this is that we add poisson noise and sky background to the ground truth in order to simulate real data taken from a telescope (see the bottom of the first cell in the Simulating a Mock Lens section in the notebook). When we perform lens modelling in lenstronomy and JAXtronomy, no noise or sky background is added to the simulated image because in a real application, we do not know what the exact noise and sky background will look like. This means that having relatively high residuals is inevitable when there is a lot of noise or sky background, and the user will often have to experiment with the tolerance level to see when they get a residual that is "good enough", which I determined to be 5200 in this specific instance.

Additionally, it is intentional that the PSO does not converge. It often takes a very long time for it to converge (on the order of hours), but the main point here is to simply compare how long it takes to run some number of iterations in lenstronomy vs JAXtronomy, so I do not bother running the PSO for too long. It also demonstrates the impressiveness of the minimizer, which is able to find a much better fit in much less time.

I will add some discussion to the notebook and experiment with more general gradient-based tolerance criteria as well.
