+++ jaxtronomy/ImSim/image_model.py
@@ -417,7 +417,7 @@ class ImageModel(object):
         :return: rendered point source images
         """
         point_source_image = jnp.zeros((self.Data.num_pixel_axes))
-        if unconvolved or self.PointSource is None:
+        if self.PointSource is None:
             return point_source_image
         ra_pos, dec_pos, amp = self.PointSource.point_source_list(
             kwargs_ps, kwargs_lens=kwargs_lens, k=k
@@ -426,7 +426,7 @@ class ImageModel(object):
             ra_pos, dec_pos, kwargs_special=kwargs_special
         )
         point_source_image += self.ImageNumerics.point_source_rendering(
-            ra_pos, dec_pos, amp
+            ra_pos, dec_pos, amp, unconvolved=unconvolved
         )
         return point_source_image * self._flux_scaling
 
