### 2026-01-14 63288f6
remove if statement
