### 2026-01-14 600cf6a
parallelize pso and mcmc
