+++ jaxtronomy/Sampling/sampler.py
@@ -1,0 +1,187 @@
+import time
+
+import numpy as np
+from jaxtronomy.Sampling.Samplers.pso import ParticleSwarmOptimizer
+from lenstronomy.Sampling.sampler import Sampler as Sampler_lenstronomy
+from lenstronomy.Util import sampling_util
+
+from functools import partial
+import jax
+from jax import lax
+
+__all__ = ["Sampler"]
+
+
+class Sampler(Sampler_lenstronomy):
+    """Inherits samplers from lenstronomy, but modifies the PSO and mcmc to parallelize
+    computations across hardware devices.
+    """
+
+    def pso(
+        self,
+        n_particles,
+        n_iterations,
+        lower_start=None,
+        upper_start=None,
+        threadCount=None,
+        init_pos=None,
+        mpi=False,
+        print_key="PSO",
+        verbose=True,
+    ):
+        """Return the best fit for the lens model on catalogue basis with particle swarm
+        optimizer.
+
+        :param n_particles: number of particles in the sampling process
+        :param n_iterations: number of iterations of the swarm
+        :param lower_start: numpy array, lower end parameter of the values of the
+            starting particles
+        :param upper_start: numpy array, upper end parameter of the values of the
+            starting particles
+        :param threadCount: number of threads in the computation (only applied if
+            mpi=False)
+        :param init_pos: numpy array, position of the initial best guess model
+        :param mpi: bool, if True, makes instance of MPIPool to allow for MPI execution
+        :param print_key: string, prints the process name in the progress bar (optional)
+        :param verbose: suppress or turn on print statements
+        :return: kwargs_result (of best fit), [lnlikelihood of samples, positions of
+            samples, velocity of samples])
+        """
+
+        if threadCount is not None:
+            raise ValueError("PSO threadCount must be None. In JAXtronomy, parallelization across devices is done automatically.")
+        if mpi:
+            raise ValueError("mpi must be False in JAXtronomy")
+
+        if lower_start is None or upper_start is None:
+            lower_start, upper_start = np.array(self.lower_limit), np.array(
+                self.upper_limit
+            )
+            print("PSO initialises its particles with default values")
+        else:
+            lower_start = np.maximum(lower_start, self.lower_limit)
+            upper_start = np.minimum(upper_start, self.upper_limit)
+
+        pso = ParticleSwarmOptimizer(
+            self.chain.logL, lower_start, upper_start, n_particles
+        )
+
+        if init_pos is None:
+            init_pos = (upper_start - lower_start) / 2 + lower_start
+
+        pso.set_global_best(init_pos, [0] * len(init_pos), self.chain.logL(init_pos))
+
+        time_start = time.time()
+
+        result, [log_likelihood_list, pos_list, vel_list] = pso.optimize(
+            n_iterations, verbose=verbose
+        )
+
+        kwargs_return = self.chain.param.args2kwargs(result)
+        if verbose:
+            print(
+                pso.global_best.fitness
+                * 2
+                / (max(self.chain.effective_num_data_points(**kwargs_return), 1)),
+                "reduced X^2 of best position",
+            )
+            print(pso.global_best.fitness, "log likelihood")
+            self._print_result(result=result)
+            time_end = time.time()
+            print(time_end - time_start, "time used for ", print_key)
+            print("===================")
+        return result, [log_likelihood_list, pos_list, vel_list]
+
+    def mcmc_emcee(
+        self,
+        n_walkers,
+        n_run,
+        n_burn,
+        mean_start,
+        sigma_start,
+        mpi=False,
+        progress=False,
+        threadCount=None,
+        initpos=None,
+        backend_filename=None,
+        start_from_backend=False,
+    ):
[…]
