### 2026-01-14 f75a013
more general tolerance criteria for minimization
