### 2026-01-14 427a537
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
