+++ jaxtronomy/Workflow/fitting_sequence.py
@@ -340,7 +340,7 @@ class FittingSequence(object):
         walkerRatio=None,
         n_walkers=None,
         sigma_scale=1,
-        threadCount=1,
+        threadCount=None,
         init_samples=None,
         re_use_samples=True,
         sampler_type="emcee",
