### 2026-01-15 #9685 @ahuang314
Hi @ConnorStoneAstro , for the `State of the Field` element in the checklist, what kind of comparisons are you looking for? Is a simple high-level overview of the differences in software functionality sufficient for the paper, or are detailed numerical comparisons between the different software for lens modeling and image simulation required?
