### 2026-01-15 871878b
parallelize only for cpu
