### 2026-01-15 375941f
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
