### 2026-01-15 2f2e13e
save logL func as instance attribute
