+++ jaxtronomy/Sampling/Samplers/pso.py
@@ -90,6 +90,8 @@ class ParticleSwarmOptimizer(PSO_lenstronomy):
 
             def logL_func(position):
                 return vmapped_func(position).flatten()
+            
+            self.logL_func = logL_func
 
     def _get_fitness(self, swarm):
         """Set fitness (probability) of the particles in swarm.
