### 2026-01-15 8542c9f
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
