### 2026-01-15 ad7e346
remove vmap
