### 2026-01-16 e72d29a
changed default rng seed to random
