### 2026-01-16 5b18662
remove rng seed argument from pso
