### 2026-01-22 #9685 @qiuhan96
@ahuang314 @ivastar Hi both, I have done the review. The software can be easily installed and run. As noted by @ConnorStoneAstro, however, running the examples also requires mannually installing tqdm, optax and numpyro via pip. Below I list a few questions and comments regarding the code and the paper.

1. The benchmark currently focuses on computing lensing quantities (e.g. deflections and surface brightness) for a set of profiles. However, there is no clear benchmark demonstrating improvements in fitting lensing images, which is a key goal of this code. I woud suggest explicitly including benchmarks for lensing image fitting performance.
2. From the benchmark records listed on the Github page, it appears that for some components, specially COVERGENCE, SHEAR and NIE, the JAX implementation are much slower than the original lenstronomy versions. For example, SHEAR appears to be about 0.7 times slower than the non-JAX implementation. Could the authors explain why this is the case? In addition, I note that benchmarks for COVERGENCE, SHEAR and NIE are not included in the benchmark table in the paper.
3. When running the tests locally, I found that EPL_numba is ~ 4 times slower than the EPL (non-jax) version. Is there any specific configuration for Numba?
4. In the Summary section of the paper, it lacks a broad description of the strong gravitational lensing phenomenon. Readers who are not familiar with lenstronomy may find it difficult to follow. I would suggest adding a brief introductory desciption of the strong lensing.
5. There is no comparison of this code to other similar tools in the community, such as the Herculens or gigalens, which are mentioned in the paper. I agree with @ConnorStoneAstro that the authors could add a few sentences explaining the strength and the motivation for using JAXtronomy.
