### 2026-01-22 #9685 @ahuang314
@qiuhan96 Thank you for your review and feedback.

1. Explicit benchmarks for the lens fitting process is not included because there are so many factors involved, notably, the mass profiles of the deflectors, light profiles of the sources, PSF kernel, convolution method, sampling method, and log likelihood function, which jaxtronomy/lenstronomy provides many options for. I think it is more useful to have the benchmarks for each individual step in the pipeline rather than for the overall process. A notebook "Modeling a simple Einstein ring" is provided for users to play around and compare between JAXtronomy's and lenstronomy's lens fitting process themselves. I can of course include a benchmark in the paper for a specific case, but I think this is not very useful for the reader. What do you think?
2. This is because in JAX, function call overheads are significantly higher than in standard Python. Thus, for functions that are already simple and fast, where JAX overheads outweigh the original computation time, JAX will be slower. I will include this explanation in the paper.
3. Not sure about this, I will have to dig into this.
4. and 5. I will include this
