### 2026-01-22 0ecfc7d
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
