### 2026-01-22 320888e
use vmap for gpu
