### 2026-01-22 c2c19ef
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
