### 2026-01-22 c1db827
handle the parallelization of logL func in Sampler
