### 2026-01-22 3606b7c
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
