### 2026-01-22 58352a1
handle parallelization of logL func in Sampler class instead
