### 2026-01-22 23e62f7
edit docstrings
