### 2026-01-22 912a600
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
