### 2026-01-22 251718a
edit raise error string
