+++ jaxtronomy/Sampling/Samplers/pso.py
@@ -3,17 +3,14 @@ from lenstronomy.Sampling.Samplers.pso import (
     Particle,
 )
 
-import jax
 import numpy as np
-from functools import partial
 
 __all__ = ["ParticleSwarmOptimizer"]
 
 
 class ParticleSwarmOptimizer(PSO_lenstronomy):
     """Optimizer using a swarm of particles. Same as the PSO from lenstronomy, but
-    parallelizes computations across CPU cores automatically using JAX. For computation
-    on GPU, only one GPU is used due to memory transfer overheads.
+    the input log likelihood function is assumed to be vectorized.
 
     :param func: A function that takes a vector in the parameter space as input and
         returns the natural logarithm of the posterior probability for that position.
