### 2026-01-23 bcd2f4d
raise error at start of function
