+++ test/test_Sampling/test_sampler.py
@@ -1,0 +1,180 @@
+__author__ = "sibirrer"
+
+import pytest
+import numpy as np, numpy.testing as npt
+import os
+import lenstronomy.Util.simulation_util as sim_util
+from lenstronomy.ImSim.image_model import ImageModel
+from jaxtronomy.Sampling.likelihood import LikelihoodModule
+from lenstronomy.Sampling.parameters import Param
+from lenstronomy.LensModel.lens_model import LensModel
+from lenstronomy.LightModel.light_model import LightModel
+from jaxtronomy.Sampling.sampler import Sampler, prepare_logL_func
+from lenstronomy.Data.imaging_data import ImageData
+from lenstronomy.Data.psf import PSF
+
+
+class TestSampler(object):
+    """Test the fitting sequences."""
+
+    def setup_method(self):
+        # data specifics
+        sigma_bkg = 0.05  # background noise per pixel
+        exp_time = 100  # exposure time (arbitrary units, flux per pixel is in units #photons/exp_time unit)
+        numPix = 10  # cutout pixel size
+        deltaPix = 0.1  # pixel size in arcsec (area per pixel = deltaPix**2)
+        fwhm = 0.5  # full width half max of PSF
+
+        # PSF specification
+
+        kwargs_data = sim_util.data_configure_simple(
+            numPix, deltaPix, exp_time, sigma_bkg
+        )
+        data_class = ImageData(**kwargs_data)
+        kwargs_psf_gaussian = {
+            "psf_type": "GAUSSIAN",
+            "fwhm": fwhm,
+            "pixel_size": deltaPix,
+        }
+        psf = PSF(**kwargs_psf_gaussian)
+        kwargs_psf = {
+            "psf_type": "PIXEL",
+            "kernel_point_source": psf.kernel_point_source,
+        }
+        psf_class = PSF(**kwargs_psf)
+        kwargs_spemd = {
+            "theta_E": 1.0,
+            "gamma": 1.8,
+            "center_x": 0,
+            "center_y": 0,
+            "e1": 0.1,
+            "e2": 0.1,
+        }
+
+        lens_model_list = ["EPL"]
+        self.kwargs_lens = [kwargs_spemd]
+        lens_model_class = LensModel(lens_model_list=lens_model_list)
+        kwargs_sersic = {
+            "amp": 1.0,
+            "R_sersic": 0.1,
+            "n_sersic": 2,
+            "center_x": 0,
+            "center_y": 0,
+        }
+        # 'SERSIC_ELLIPSE': elliptical Sersic profile
+        kwargs_sersic_ellipse = {
+            "amp": 1.0,
+            "R_sersic": 0.6,
+            "n_sersic": 3,
+            "center_x": 0,
+            "center_y": 0,
+            "e1": 0.1,
+            "e2": 0.1,
+        }
+
+        lens_light_model_list = ["SERSIC"]
+        self.kwargs_lens_light = [kwargs_sersic]
+        lens_light_model_class = LightModel(light_model_list=lens_light_model_list)
+        source_model_list = ["SERSIC_ELLIPSE"]
+        self.kwargs_source = [kwargs_sersic_ellipse]
+        source_model_class = LightModel(light_model_list=source_model_list)
+
+        kwargs_numerics = {
+            "supersampling_factor": 1,
+            "supersampling_convolution": False,
+            "compute_mode": "regular",
+        }
+        imageModel = ImageModel(
+            data_class,
+            psf_class,
+            lens_model_class,
+            source_model_class,
+            lens_light_model_class,
+            kwargs_numerics=kwargs_numerics,
+        )
+        image_sim = sim_util.simulate_simple(
[…]
