+++ test/test_Workflow/test_fitting_sequence.py
@@ -1194,11 +1194,10 @@ class TestFittingSequence(object):
         )
 
         optax_kwargs = {
-            "maxiter": 300,
+            "maxiter": 1000,
             "num_chains": 1,
-            "tolerance": 100,
+            "tolerance": 0.01,
             "sigma_scale": 1,
-            "rng_seed": 1,
         }
         fitting_kwargs_list_jaxopt = [["optax", optax_kwargs]]
         chain_list = fitting_seq.fit_sequence(fitting_kwargs_list_jaxopt)
