### 2026-01-23 3159461
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
