### 2026-01-23 2bde497
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
