### 2026-01-23 a2a0759
remove unused return value
