### 2026-01-25 #9685 @qiuhan96
Hi @ahuang314 , thanks! As for the first point, would it be possible to do a comparison that by fitting images with exactly the same deflectors, source models, PSFs using JAX and non-JAX version? Yes, I have run the fitting example, and it is super impressive that how fast the fitting is with optax.
