### 2026-02-03 9cd9b76
copy lenstronomy PR 812
