+++ jaxtronomy/Sampling/Likelihoods/position_likelihood.py
@@ -20,7 +20,7 @@ class PositionLikelihood(object):
         image_position_likelihood=False,
         ra_image_list=None,
         dec_image_list=None,
-        source_position_likelihood=False,
+        source_position_likelihood=None,
         source_position_tolerance=None,
         source_position_sigma=0.001,
         force_no_add_image=False,
@@ -56,6 +56,14 @@ class PositionLikelihood(object):
         self._pointSource = point_source_class
         # TODO replace with public function of ray_shooting
         self._lensModel = point_source_class._lens_model
+
+        point_source_type_list = point_source_class.point_source_type_list
+
+        if source_position_likelihood is None:
+            if "LENSED_POSITION" in point_source_type_list:
+                source_position_likelihood = True
+            else:
+                source_position_likelihood = False
 
         # TODO: Implement restrict image number and force no add image
         if force_no_add_image:
