### 2026-02-13 #9685 @ahuang314
Hi @ConnorStoneAstro @qiuhan96,

I have addressed all of your comments:

- Dependencies on PyPI fixed ([issue #70](https://github.com/lenstronomy/JAXtronomy/issues/70))
- readme uploaded to jaxtronomy's PyPI page ([issue #71](https://github.com/lenstronomy/JAXtronomy/issues/71))
- PSO and MCMC now compute the likelihood function for particles/walkers in parallel to take full advantage of GPU architecture ([issue #72](https://github.com/lenstronomy/JAXtronomy/issues/72))
- gradient descent tolerance criteria has been generalized so that iterations stop whenever the following criteria is met three times in a row: `| logL(i) - logL(i - 1) | < tolerance` ([issue #73](https://github.com/lenstronomy/JAXtronomy/issues/73))
- Paper now includes specific benchmarks between jaxtronomy and lenstronomy for a specific lens fitting example (qiuhan96's first comment)
- Paper now includes explanation about why certain mass profiles are computed slower in JAXtronomy than in lenstronomy, and those profiles have been added to the table (qiuhan96's second comment)
- EPL_numba is not actually 4x slower than EPL (non-jax). The comparison notebook had an error where the number of for-loop iterations was set to 10000 for EPL_numba and 1000 for everything else, resulting in an erroneous comparison. The notebook has been fixed (qiuhan96's third comment)
- Summary section of paper now includes a very brief explanation of strong lensing (qiuhan96's fourth comment)
- Paper now includes caustics citation. Rather than directly comparing jaxtronomy to each of the individually mentioned JAX/GPU-accelerated strong lensing software packages, I have instead chosen to point out a few of JAXtronomy's features that are not included in _any_ of the other packages as justification for why JAXtronomy is a useful addition to the field. I hope that this is a sufficient alternative.
