### 2026-02-14 #9685 @ConnorStoneAstro
These are excellent updates @ahuang314 ! I went to your "paper" branch to see the updated paper, you may want to compile a new draft here too. I see in the PSO comparison that the GPU version achieves 46x speedup over lenstronomy with 512 particles, is that with the new vectorized PSO?

Either way, it looks like you've addressed all of my comments and I am really happy with the state things are in. JAXtronomy is a great boost for a lot of projects! @ivastar I am happy to recommend publication.
