### 2026-02-14 #9685 @ahuang314
Yes, the 46x speedup is with the new vectorized PSO, and the paper pdf you saw was already the latest version. I tried to reproduce the 1000x speedup you saw with caustics, but the best I could find was a 100x speedup.
