### 2026-02-14 #9685 @ConnorStoneAstro
@ahuang314 , to get the 1000x speedup we selected some parameters that would slightly favour a GPU. We had a 128x128 grid which we 4x upsampled, so really it was a 512x512 grid, plus we did PSF convolution, then we reached the 1000x speedup when doing a batch of ~100 of these simultaneously (this is actually pretty close to what we were doing for our research, so not an unreasonable choice). We also didn't include tracing in the timing calculation.
