### 2026-02-26 248affe
Add macOS GPU (Metal) runtime config and Metal GPU support
Introduce a runtime config helper to detect macOS+Metal and set jax_enable_x64 accordingly (jaxtronomy/_runtime_config.py). Replace ad-hoc jax.config.update(...) calls with configure_jax_precision_for_runtime() across the codebase. Add Metal workarounds: a Metal-safe real-valued EPL series implementation and variants (function_/derivatives_/hessian_real_series) plus spatial (lax.conv) convolution fallback when FFT legalization fails on Metal. Normalize backend labels and add Metal-specific input sanitization/clip logic in the sampler a
[…]
