+++ jaxtronomy/LensModel/single_plane.py
@@ -1,11 +1,11 @@
 __author__ = "sibirrer"
 
 import jax
+from jaxtronomy._runtime_config import configure_jax_precision_for_runtime
 from jax import jit, numpy as jnp
 from jaxtronomy.LensModel.profile_list_base import ProfileListBase
 from functools import partial
-
-jax.config.update("jax_enable_x64", True)
+configure_jax_precision_for_runtime()
 
 __all__ = ["SinglePlane"]
 
