+++ jaxtronomy/Sampling/likelihood.py
@@ -2,8 +2,8 @@ __author__ = "sibirrer"
 
 from functools import partial
 import jax
-
-jax.config.update("jax_enable_x64", True)
+from jaxtronomy._runtime_config import configure_jax_precision_for_runtime
+configure_jax_precision_for_runtime()
 from jax import jit, lax, numpy as jnp
 import numpy as np
 
