+++ jaxtronomy/Sampling/Samplers/optax.py
@@ -1,8 +1,8 @@
 __author__ = "ahuang314"
 
 import jax
-
-jax.config.update("jax_enable_x64", True)
+from jaxtronomy._runtime_config import configure_jax_precision_for_runtime
+configure_jax_precision_for_runtime()
 from jax import jit, numpy as jnp
 from functools import partial
 import optax
