+++ jaxtronomy/LensModel/MultiPlane/decoupled_multi_plane.py
@@ -2,8 +2,8 @@ __author__ = "dangilman"
 
 from functools import partial
 from jax import config, jit
-
-config.update("jax_enable_x64", True)
+from jaxtronomy._runtime_config import configure_jax_precision_for_runtime
+configure_jax_precision_for_runtime()
 
 from lenstronomy.LensModel.MultiPlane.multi_plane import MultiPlane
 from lenstronomy.Cosmo.background import Background
