+++ test/test_LensModel/test_Profiles/test_epl_gpu_cpu_parity.py
@@ -1,0 +1,160 @@
+"""GPU-vs-CPU parity checks for EPL in jaxtronomy."""
+
+from __future__ import annotations
+
+import json
+import os
+import subprocess
+import sys
+from typing import Dict, Optional
+
+import numpy as np
+import pytest
+
+
+def _json_from_subprocess_output(stdout: str, stderr: str) -> Dict[str, object]:
+    lines = [line.strip() for line in stdout.splitlines() if line.strip()]
+    lines.extend(line.strip() for line in stderr.splitlines() if line.strip())
+    for line in reversed(lines):
+        try:
+            return json.loads(line)
+        except Exception:
+            continue
+    raise RuntimeError("No JSON payload received from subprocess.")
+
+
+def _probe_accelerator_platform() -> Optional[str]:
+    code = r"""
+import json
+import jax
+
+platforms = [str(getattr(d, "platform", "")).strip().lower() for d in jax.devices()]
+accel = next((p for p in platforms if p not in {"", "cpu"}), None)
+print(json.dumps({"platforms": platforms, "accelerator": accel}))
+"""
+    env = os.environ.copy()
+    env.pop("JAX_PLATFORM_NAME", None)
+    proc = subprocess.run(
+        [sys.executable, "-c", code],
+        check=True,
+        capture_output=True,
+        text=True,
+        env=env,
+    )
+    payload = _json_from_subprocess_output(proc.stdout, proc.stderr)
+    accel = payload.get("accelerator")
+    if accel is None:
+        return None
+    accel_s = str(accel).strip().lower()
+    return accel_s or None
+
+
+def _run_epl_payload(platform_env: Optional[str]) -> Dict[str, object]:
+    code = r"""
+import json
+import numpy as np
+import jax
+from jaxtronomy.LensModel.Profiles.epl import EPL
+
+rng = np.random.default_rng(20260226)
+x = rng.uniform(-2.0, 2.0, size=128).astype(np.float32)
+y = rng.uniform(-2.0, 2.0, size=128).astype(np.float32)
+
+# Stay away from the singular center to keep parity metrics stable.
+r = np.hypot(x, y)
+mask = r < 0.15
+x[mask] = x[mask] + 0.35
+y[mask] = y[mask] - 0.25
+
+theta_E = 1.15
+gamma = 2.08
+e1 = 0.10
+e2 = -0.07
+
+profile = EPL()
+
+f = np.asarray(profile.function(x, y, theta_E, gamma, e1, e2), dtype=np.float64)
+fx, fy = profile.derivatives(x, y, theta_E, gamma, e1, e2)
+fx = np.asarray(fx, dtype=np.float64)
+fy = np.asarray(fy, dtype=np.float64)
+fxx, fxy, fyx, fyy = profile.hessian(x, y, theta_E, gamma, e1, e2)
+fxx = np.asarray(fxx, dtype=np.float64)
+fxy = np.asarray(fxy, dtype=np.float64)
+fyx = np.asarray(fyx, dtype=np.float64)
+fyy = np.asarray(fyy, dtype=np.float64)
+
+payload = {
+    "default_backend": str(jax.default_backend()).strip().lower(),
+    "device_platform": str(getattr(jax.devices()[0], "platform", "")).strip().lower(),
+    "function": f.tolist(),
+    "fx": fx.tolist(),
+    "fy": fy.tolist(),
+    "fxx": fxx.tolist(),
+    "fxy": fxy.tolist(),
+    "fyx": fyx.tolist(),
+    "fyy": fyy.tolist(),
+}
+print(json.dumps(payload))
+"""
+    env = os.environ.copy()
[…]
