+++ jaxtronomy/LensModel/Profiles/null.py
@@ -1,9 +1,9 @@
 import jax
+from jaxtronomy._runtime_config import configure_jax_precision_for_runtime
 from jax import jit, numpy as jnp
 
 from lenstronomy.LensModel.Profiles.base_profile import LensProfileBase
-
-jax.config.update("jax_enable_x64", True)
+configure_jax_precision_for_runtime()
 
 __all__ = ["Null"]
 
