### 2026-02-26 1490349
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
