+++ jaxtronomy/LensModel/Profiles/nfw.py
@@ -7,6 +7,7 @@ from jaxtronomy._runtime_config import configure_jax_precision_for_runtime
 import jax.numpy as jnp
 
 from lenstronomy.LensModel.Profiles.base_profile import LensProfileBase
+
 configure_jax_precision_for_runtime()
 
 __all__ = ["NFW"]
