+++ jaxtronomy/Sampling/likelihood.py
@@ -3,6 +3,7 @@ __author__ = "sibirrer"
 from functools import partial
 import jax
 from jaxtronomy._runtime_config import configure_jax_precision_for_runtime
+
 configure_jax_precision_for_runtime()
 from jax import jit, lax, numpy as jnp
 import numpy as np
