+++ jaxtronomy/Sampling/Samplers/optax.py
@@ -2,6 +2,7 @@ __author__ = "ahuang314"
 
 import jax
 from jaxtronomy._runtime_config import configure_jax_precision_for_runtime
+
 configure_jax_precision_for_runtime()
 from jax import jit, numpy as jnp
 from functools import partial
