+++ jaxtronomy/LensModel/Profiles/epl_multipole_m3m4.py
@@ -2,6 +2,7 @@ __author__ = "dangilman"
 
 from jax import config, jit, numpy as jnp
 from jaxtronomy._runtime_config import configure_jax_precision_for_runtime
+
 configure_jax_precision_for_runtime()
 
 from lenstronomy.LensModel.Profiles.base_profile import LensProfileBase
