### 2026-03-06 0f6fc9a
fix docstrings
