### 2026-03-06 996f912
add tests
