+++ test/test_ImSim/test_MultiBand/test_multi_linear.py
@@ -119,12 +119,14 @@ class TestImageModel(object):
             [kwargs_data, kwargs_psf, kwargs_numerics],
             [kwargs_data, kwargs_psf, kwargs_numerics],
         ]
+        self.multi_band_list = multi_band_list
         kwargs_model = {
             "lens_model_list": lens_model_list,
             "source_light_model_list": source_model_list,
             "point_source_model_list": ["LENSED_POSITION"],
             "fixed_magnification_list": [True],
         }
+        self.kwargs_model = kwargs_model
         self.imageModel = MultiLinear(
             multi_band_list,
             kwargs_model,
@@ -137,6 +139,21 @@ class TestImageModel(object):
             likelihood_mask_list=None,
             compute_bool=[True, False],
         )
+
+    def test_init(self):
+        npt.assert_raises(ValueError,
+            MultiLinear,
+            self.multi_band_list,
+            self.kwargs_model,
+            likelihood_mask_list=None,
+            compute_bool=[True, False, False],
+        )
+        assert self.imageModel.num_bands == self.imageModel_ref.num_bands
+        assert self.imageModel.num_response_list == self.imageModel_ref.num_response_list
+        assert self.imageModel.num_data_evaluate == self.imageModel_ref.num_data_evaluate
+        num_param_linear = self.imageModel.num_param_linear(self.kwargs_lens, self.kwargs_source, self.kwargs_lens_light, self.kwargs_ps)
+        num_param_linear_ref = self.imageModel_ref.num_param_linear(self.kwargs_lens, self.kwargs_source, self.kwargs_lens_light, self.kwargs_ps)
+        assert num_param_linear == num_param_linear_ref
 
     def test_image_linear_solve(self):
         model, error_map, cov_param, param = self.imageModel.image_linear_solve(
@@ -187,6 +204,13 @@ class TestImageModel(object):
             else:
                 npt.assert_allclose(param[i], param_ref[i], atol=1e-6, rtol=1e-6)
 
+        residuals = self.imageModel.reduced_residuals(model, error_map)
+        residuals_ref = self.imageModel_ref.reduced_residuals(model_ref, error_map_ref)
+        assert len(residuals) == len(residuals_ref)
+        for i in range(len(residuals)):
+            npt.assert_allclose(residuals[i], residuals_ref[i], atol=1e-5, rtol=1e-5)
+            
+
     def test_likelihood_data_given_model(self):
         logL, param = self.imageModel.likelihood_data_given_model(
             self.kwargs_lens,
@@ -212,11 +236,26 @@ class TestImageModel(object):
             else:
                 npt.assert_allclose(param[i], param_ref[i], atol=1e-6, rtol=1e-6)
 
-    def test_numData_evaluate(self):
-        numData = self.imageModel.num_data_evaluate
-        numData_ref = self.imageModel_ref.num_data_evaluate
-        assert numData == numData_ref
-
+    def test_update_linear_kwargs(self):
+        num_param_linear = self.imageModel.num_param_linear(self.kwargs_lens, self.kwargs_source, self.kwargs_lens_light, self.kwargs_ps)
+        param = np.ones(int(num_param_linear)) * 10
+        kwargs = self.imageModel.update_linear_kwargs(
+            [param*2, param],
+            0,
+            self.kwargs_lens,
+            self.kwargs_source,
+            self.kwargs_lens_light,
+            self.kwargs_ps,
+        )
+        kwargs_ref = self.imageModel_ref.update_linear_kwargs(
+            [param*2, param],
+            0,
+            self.kwargs_lens,
+            self.kwargs_source,
+            self.kwargs_lens_light,
+            self.kwargs_ps,
+        )
+        assert kwargs_ref == kwargs
 
 if __name__ == "__main__":
     pytest.main()
