### 2026-03-07 de7c6bc
loosen test assertion
