### 2026-03-07 730ae44
array of zeros instead of list
