+++ jaxtronomy/ImSim/MultiBand/multi_data_base.py
@@ -83,7 +83,7 @@ class MultiDataBase(object):
         """
         residual_list = []
         if error_map_list is None:
-            error_map_list = [[] for _ in range(self._num_bands)]
+            error_map_list = jnp.zeros(self._num_bands)
         index = 0
         for i in range(self._num_bands):
             if self._compute_bool[i] is True:
