### 2026-03-16 81c4871
[pre-commit.ci] pre-commit autoupdate
updates:
- https://github.com/psf/black → https://github.com/psf/black-pre-commit-mirror
- [github.com/psf/black-pre-commit-mirror: 25.1.0 → 26.3.1](https://github.com/psf/black-pre-commit-mirror/compare/25.1.0...26.3.1)
- https://github.com/psf/black → https://github.com/psf/black-pre-commit-mirror
- [github.com/psf/black-pre-commit-mirror: 25.1.0 → 26.3.1](https://github.com/psf/black-pre-commit-mirror/compare/25.1.0...26.3.1)
