### 2026-03-30 #9685 @ivastar
@ahuang314 I have a couple of minor editorial comments. Please let me know if you have questions. Ping me when you've had a chance to make these changes.

Line 15-17: "Studying graviationally lensed objects can give us key insights into cosmology and astrophysics, such as constraints on the expansion rate of the universe and dark matter models." --> there is a typo in "gravitationally" but also the sentence sounds a bit off, perhaps "Studying gravitationally lensed systems provides insights into cosmology and astrophysics, including constraints of the expansion rate of the Universe and the distribution of dark matter."

Line 19: no need to add a footnote to the github repo, there is a proper citation for `lenstronomy`. Same for `caustics` on line 51 and for `jaxtstronomy`. For the latter, include a citation:
@software{jaxstronomy,
  author  = {{Google Research}},
  title   = {jaxstronomy: JAX-based tools for astronomical inference},
  year    = {2023},
  url     = {https://github.com/google-research/google-research/tree/master/jaxstronomy}
}

Also for `optax` on line 113.

Line 19: Instead of including JAX in the footnotes, please add a citation:
@software{jax2018github,
  author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
  title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
  url = {http://github.com/jax-ml/jax},
  version = {0.3.13},
  year = {2018},
}

Line 36: the TDCOSMO2025 citation is not parsed correctly

Line 55: Such --> These

Line 110-111: Remove "These benchmarks were performed...." The previous section already makes it clear that the benchmarks were all performed on the same hardware.

Overall, there is something wonky with the citations - some display full names, some just first initials, some nothing at al. I would go through the bib file and make sure these are done in a standard way.

A specific comment about the reference on line 134 for which the first author is "Collaboration".

Please check if there are DOI citations for the missing DOIs above.

And finally, please add to the acknowledgements citations for the major dependancies of JAXtronomy that have not been already mentioned in the text.
