+++ paper/paper.bib
@@ -27,6 +27,14 @@ journal = {The Astrophysical Journal},
 abstract = {Gravitationally lensed curved arcs provide a wealth of information about the underlying lensing distortions. Extracting precise lensing information from extended sources is a key component in many studies aiming to answer fundamental questions about the universe. To maintain accuracy with increased precision, it is of vital importance to characterize and understand the impact of degeneracies inherent in lensing observables. In this work, we present a formalism to describe the gravitational lensing distortion effects resulting in curved extended arcs based on the eigenvectors and eigenvalues of the local lensing Jacobian and their directional differentials. We identify a nonlocal and nonlinear extended deflector basis that inherits these local properties. Our parameterization is tightly linked to observable features in extended sources and allows one to accurately extract the lensing information of extended images without imposing an explicit global deflector model. We quantify what degeneracies can be broken based on specific assumptions about the local lensing nature and assumed intrinsic source shape. Our formalism is applicable from the weak linear regime to the semi-linear regime and all the way up to the highly nonlinear regime of highly magnified arcs of multiple images. The methodology and implementation presented in this work provides a framework to assessing systematics, to guide inference efforts in the right choices in complexity based on the data at hand, and to quantify the lensing information extracted in a model-independent way (https://github.com/sibirrer/curved_arcs).}
 }
 
+@software{JAX,
+  author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
+  title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
+  url = {http://github.com/jax-ml/jax},
+  version = {0.3.13},
+  year = {2018},
+}
+
 @ARTICLE{Shajib:2021,
        author = {{Shajib}, Anowar J. and {Treu}, Tommaso and {Birrer}, Simon and {Sonnenfeld}, Alessandro},
         title = "{Dark matter haloes of massive elliptical galaxies at z {\ensuremath{\sim}} 0.2 are well described by the Navarro-Frenk-White profile}",
@@ -44,8 +52,6 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2021MNRAS.503.2380S},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
-
-
 
 @ARTICLE{DINOS1,
        author = {{Tan}, Chin Yi and {Shajib}, Anowar J. and {Birrer}, Simon and {Sonnenfeld}, Alessandro and {Treu}, Tommaso and {Wells}, Patrick and {Williams}, Devon and {Buckley-Geer}, Elizabeth J. and {Drlica-Wagner}, Alex and {Frieman}, Joshua},
@@ -83,15 +89,19 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-#TODO: Replace with journal BibTeX
-@misc{Shajib:2025,
-    author = "Shajib, Anowar J. and Nihal, Nafis Sadik and Tan, Chin Yi and Sahu, Vedant and Birrer, Simon and Treu, Tommaso and Frieman, Joshua",
-    title = "{dolphin: A fully automated forward modeling pipeline powered by artificial intelligence for galaxy-scale strong lenses}",
-    eprint = "2503.22657",
-    archivePrefix = "arXiv",
-    primaryClass = "astro-ph.IM",
-    month = "3",
-    year = "2025"
+@article{Shajib:2025,
[…]
