+++ paper/paper.bib
@@ -21,7 +21,7 @@ publisher = {The American Astronomical Society},
 volume = {919},
 number = {1},
 pages = {38},
-author = {Birrer, Simon},
+author = {{Birrer}, Simon},
 title = {Gravitational Lensing Formalism in a Curved Arc Basis: A Continuous Description of Observables and Degeneracies from the Weak to the Strong Lensing Regime},
 journal = {The Astrophysical Journal},
 abstract = {Gravitationally lensed curved arcs provide a wealth of information about the underlying lensing distortions. Extracting precise lensing information from extended sources is a key component in many studies aiming to answer fundamental questions about the universe. To maintain accuracy with increased precision, it is of vital importance to characterize and understand the impact of degeneracies inherent in lensing observables. In this work, we present a formalism to describe the gravitational lensing distortion effects resulting in curved extended arcs based on the eigenvectors and eigenvalues of the local lensing Jacobian and their directional differentials. We identify a nonlocal and nonlinear extended deflector basis that inherits these local properties. Our parameterization is tightly linked to observable features in extended sources and allows one to accurately extract the lensing information of extended images without imposing an explicit global deflector model. We quantify what degeneracies can be broken based on specific assumptions about the local lensing nature and assumed intrinsic source shape. Our formalism is applicable from the weak linear regime to the semi-linear regime and all the way up to the highly nonlinear regime of highly magnified arcs of multiple images. The methodology and implementation presented in this work provides a framework to assessing systematics, to guide inference efforts in the right choices in complexity based on the data at hand, and to quantify the lensing information extracted in a model-independent way (https://github.com/sibirrer/curved_arcs).}
@@ -98,7 +98,7 @@ archivePrefix = {arXiv},
   volume = {992},
   number = {1},
   pages = {40},
-  author = {Shajib, Anowar J. and Nihal, Nafis Sadik and Tan, Chin Yi and Sahu, Vedant and Birrer, Simon and Treu, Tommaso and Frieman, Joshua},
+  author = {{Shajib}, Anowar J. and {Nihal}, Nafis Sadik and {Tan}, Chin Yi and {Sahu}, Vedant and {Birrer}, Simon and {Treu}, Tommaso and {Frieman}, Joshua},
   title = {dolphin: A Fully Automated Forward-modeling Pipeline Powered by Artificial Intelligence for Galaxy-scale Strong Lenses},
   journal = {The Astrophysical Journal},
[…]
