### 2026-03-31 e5bda89
fix bibtex
