+++ paper/paper.bib
@@ -7,7 +7,7 @@ year = {2018},
 issn = {2212-6864},
 doi = {10.1016/j.dark.2018.11.002},
 url = {https://www.sciencedirect.com/science/article/pii/S2212686418301869},
-author = {Simon Birrer and Adam Amara},
+author = {{Birrer}, Simon and {Amara}, Adam},
 keywords = {Gravitational lensing, Software, Image simulations},
 abstract = {We present lenstronomy, a multi-purpose open-source gravitational lens modelling pythonpackage. lenstronomy is able to reconstruct the lens mass and surface brightness distributions of strong lensing systems using forward modelling. lenstronomy supports a wide range of analytic lens and light models in arbitrary combination. The software is also able to reconstruct complex extended sources (Birrer et. al 2015) as well as being able to model point sources. We designed lenstronomy to be stable, flexible and numerically accurate, with a clear user interface that could be deployed across different platforms. Throughout its development, we have actively used lenstronomy to make several measurements including deriving constraints on dark matter properties in strong lenses, measuring the expansion history of the universe with time-delay cosmography, measuring cosmic shear with Einstein rings and decomposing quasar and host galaxy light. The software is distributed under the MIT license. The documentation, starter guide, example notebooks, source code and installation guidelines can be found at https://lenstronomy.readthedocs.io.}
 }
@@ -21,7 +21,7 @@ publisher = {The American Astronomical Society},
 volume = {919},
 number = {1},
 pages = {38},
-author = {{Birrer}, Simon},
+author = {Simon Birrer},
 title = {Gravitational Lensing Formalism in a Curved Arc Basis: A Continuous Description of Observables and Degeneracies from the Weak to the Strong Lensing Regime},
 journal = {The Astrophysical Journal},
 abstract = {Gravitationally lensed curved arcs provide a wealth of information about the underlying lensing distortions. Extracting precise lensing information from extended sources is a key component in many studies aiming to answer fundamental questions about the universe. To maintain accuracy with increased precision, it is of vital importance to characterize and understand the impact of degeneracies inherent in lensing observables. In this work, we present a formalism to describe the gravitational lensing distortion effects resulting in curved extended arcs based on the eigenvectors and eigenvalues of the local lensing Jacobian and their directional differentials. We identify a nonlocal and nonlinear extended deflector basis that inherits these local properties. Our parameterization is tightly linked to observable features in extended sources and allows one to accurately extract the lensing information of extended images without imposing an explicit global deflector model. We quantify what degeneracies can be broken based on specific assumptions about the local lensing nature and assumed intrinsic source shape. Our formalism is applicable from the weak linear regime to the semi-linear regime and all the way up to the highly nonlinear regime of highly magnified arcs of multiple images. The methodology and implementation presented in this work provides a framework to assessing systematics, to guide inference efforts in the right choices in complexity based on the data at hand, and to quantify the lensing information extracted in a model-independent way (https://github.com/sibirrer/curved_arcs).}
@@ -219,7 +219,7 @@ archivePrefix = {arXiv},
 }
 
 @article{Gilman:2019,
-    author = {Gilman, Daniel and Birrer, Simon and Nierenberg, Anna and Treu, Tommaso and Du, Xiaolong and Benson, Andrew},
+    author = {{Gilman}, Daniel and {Birrer}, Simon and {Nierenberg}, Anna and {Treu}, Tommaso and {Du}, Xiaolong and {Benson}, Andrew},
     title = {Warm dark matter chills out: constraints on the halo mass function and the free-streaming length of dark matter with eight quadruple-image strong gravitational lenses},
     journal = {Monthly Notices of the Royal Astronomical Society},
     volume = {491},
