+++ paper/paper.bib
@@ -7,7 +7,7 @@ year = {2018},
 issn = {2212-6864},
 doi = {10.1016/j.dark.2018.11.002},
 url = {https://www.sciencedirect.com/science/article/pii/S2212686418301869},
-author = {{Birrer}, Simon and {Amara}, Adam},
+author = {Birrer, Simon and Amara, Adam},
 keywords = {Gravitational lensing, Software, Image simulations},
 abstract = {We present lenstronomy, a multi-purpose open-source gravitational lens modelling pythonpackage. lenstronomy is able to reconstruct the lens mass and surface brightness distributions of strong lensing systems using forward modelling. lenstronomy supports a wide range of analytic lens and light models in arbitrary combination. The software is also able to reconstruct complex extended sources (Birrer et. al 2015) as well as being able to model point sources. We designed lenstronomy to be stable, flexible and numerically accurate, with a clear user interface that could be deployed across different platforms. Throughout its development, we have actively used lenstronomy to make several measurements including deriving constraints on dark matter properties in strong lenses, measuring the expansion history of the universe with time-delay cosmography, measuring cosmic shear with Einstein rings and decomposing quasar and host galaxy light. The software is distributed under the MIT license. The documentation, starter guide, example notebooks, source code and installation guidelines can be found at https://lenstronomy.readthedocs.io.}
 }
@@ -36,7 +36,7 @@ abstract = {Gravitationally lensed curved arcs provide a wealth of information a
 }
 
 @ARTICLE{Shajib:2021,
-       author = {{Shajib}, Anowar J. and {Treu}, Tommaso and {Birrer}, Simon and {Sonnenfeld}, Alessandro},
+       author = {Shajib, Anowar J. and Treu, Tommaso and Birrer, Simon and Sonnenfeld, Alessandro},
         title = "{Dark matter haloes of massive elliptical galaxies at z {\ensuremath{\sim}} 0.2 are well described by the Navarro-Frenk-White profile}",
       journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {gravitational lensing: strong, galaxies: elliptical and lenticular, cD, Astrophysics - Astrophysics of Galaxies},
