+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@article{Birrer:2018,
+@ARTICLE{Birrer:2018,
 title = {lenstronomy: Multi-purpose gravitational lens modelling software package},
 journal = {Physics of the Dark Universe},
 volume = {22},
@@ -7,12 +7,12 @@ year = {2018},
 issn = {2212-6864},
 doi = {10.1016/j.dark.2018.11.002},
 url = {https://www.sciencedirect.com/science/article/pii/S2212686418301869},
-author = {Birrer, Simon and Amara, Adam},
+author = {{Birrer}, Simon and {Amara}, Adam},
 keywords = {Gravitational lensing, Software, Image simulations},
 abstract = {We present lenstronomy, a multi-purpose open-source gravitational lens modelling pythonpackage. lenstronomy is able to reconstruct the lens mass and surface brightness distributions of strong lensing systems using forward modelling. lenstronomy supports a wide range of analytic lens and light models in arbitrary combination. The software is also able to reconstruct complex extended sources (Birrer et. al 2015) as well as being able to model point sources. We designed lenstronomy to be stable, flexible and numerically accurate, with a clear user interface that could be deployed across different platforms. Throughout its development, we have actively used lenstronomy to make several measurements including deriving constraints on dark matter properties in strong lenses, measuring the expansion history of the universe with time-delay cosmography, measuring cosmic shear with Einstein rings and decomposing quasar and host galaxy light. The software is distributed under the MIT license. The documentation, starter guide, example notebooks, source code and installation guidelines can be found at https://lenstronomy.readthedocs.io.}
 }
 
-@article{Birrer:2021,
+@ARTICLE{Birrer:2021,
 doi = {10.3847/1538-4357/ac1108},
 url = {https://dx.doi.org/10.3847/1538-4357/ac1108},
 year = {2021},
@@ -21,7 +21,7 @@ publisher = {The American Astronomical Society},
 volume = {919},
 number = {1},
 pages = {38},
-author = {Simon Birrer},
+author = {{Birrer}, Simon},
 title = {Gravitational Lensing Formalism in a Curved Arc Basis: A Continuous Description of Observables and Degeneracies from the Weak to the Strong Lensing Regime},
 journal = {The Astrophysical Journal},
[…]
