+++ paper/paper.bib
@@ -265,7 +265,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Galan:2022,
-       author = {Georgios Vernardos, Austin Peel, Frédéric Courbin, Jean-Luc Starck},
+       author = {{Vernardos}, Georgios and {Peel}, Austin and {Courbin}, Frédéric and {Starck}, Jean-Luc},
         title = "{Using wavelets to capture deviations from smoothness in galaxy-scale strong lenses}",
       journal = {Astronomy & Astrophysics},
      keywords = {galaxies: structure, dark matter, methods: data analysis, gravitation, gravitational lensing: strong, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
