+++ paper/paper.md
@@ -148,4 +148,8 @@ In the following table, we compare `JAXtronomy`'s PSO performance to that of `le
 
 Additionally, using `JAX`'s autodifferentiation, we have implemented the L-BFGS gradient descent algorithm from the `Optax` library [@DeepMind:2020] for optimization. This is a significant improvement over `lenstronomy`'s PSO, which does not have access to gradient information. Due to the stochastic nature of the PSO, we do not present a concrete comparison between `lenstronomy`'s PSO and `JAXtronomy`'s minimizer for the time it takes to find best-fit parameters.
 
+# Acknowledgements
+
+AH and SB are supported by DoE Grant DE-SC0026113, NASA Grants JWST-GO-07184 and 22-ROMAN22-0072. Major software dependencies of JAXtronomy not previously mentioned include `numpy` [@numpy], `scipy` [@scipy], and `numpyro` [@numpyro1, @numpyro2].
+
 # References
