### 2026-03-31 #9685 @ahuang314
Thank you for the comments @ivastar. I have addressed all of them. I have also modified the paragraph on line 80.

There should now be five missing DOIs: one for each of the software citations for `JAX`, `jaxstronomy`, and `optax`, as well as two from `numpyro` citations which I've added to the acknowledgements.
