### 2026-03-31 #9685 @ivastar
@ahuang314 please go ahead with the remainder of the post-review checklist. Specifically, please make sure that the Zenodo archive has the same title, author list and license and the JOSS paper.
