### 2026-04-14 #9685 @ahuang314
@ivastar I've just uploaded version 0.1.3 of JAXtronomy onto Zenodo. The DOI is https://doi.org/10.5281/zenodo.19562004
