### 2026-04-14 #9685 @ivastar
@ahuang314 the Zenodi archive lists a Creative Commons license but the actual code is BSD. Can you please update that?
