### 2026-04-14 #9685 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
