### 2026-04-14 #9685 @ivastar
@ahuang314 everything looks good! The Editor in Chief will now review the submission and do the final publication magic.

@ConnorStoneAstro, @qiuhan96 thank you for you reviewing this submission! These efforts make JOSS possible.
