### 2026-04-15 f92c193
add missing models to _JAXXED_MODELS
