### 2026-04-15 691ab18
add models to giant if statement
