### 2026-04-17 a434a30
change defaults, improve various error messages and documentation
