+++ jaxtronomy/Sampling/sampler.py
@@ -22,7 +22,7 @@ class Sampler(Sampler_lenstronomy):
         n_iterations,
         lower_start=None,
         upper_start=None,
-        threadCount=None,
+        threadCount=1,
         init_pos=None,
         mpi=False,
         print_key="PSO",
@@ -37,32 +37,30 @@ class Sampler(Sampler_lenstronomy):
             starting particles
         :param upper_start: numpy array, upper end parameter of the values of the
             starting particles
-        :param threadCount: number of threads in the computation (only applied if
-            mpi=False)
+        :param threadCount: number of threads in the computation, only relevant for CPU parallelization
         :param init_pos: numpy array, position of the initial best guess model
-        :param mpi: bool, if True, makes instance of MPIPool to allow for MPI execution
+        :param mpi: bool, if True, makes instance of MPIPool to allow for MPI execution (must be False in JAXtronomy)
         :param print_key: string, prints the process name in the progress bar (optional)
         :param verbose: suppress or turn on print statements
         :return: kwargs_result (of best fit), [lnlikelihood of samples, positions of
             samples, velocity of samples])
         """
-
-        if threadCount is not None:
+        if threadCount > len(jax.devices()):
             raise ValueError(
-                "PSO threadCount must be None. In JAXtronomy, parallelization across CPU cores is done automatically."
+                f"Supplied threadCount {threadCount} is greater than {len(jax.devices())}, the number of devices detectable by JAX.\n"
+                f"To ensure that the correct number of devices is recognized by JAX, an environment variable must be set; see JAX or JAXtronomy documentation."
             )
         if mpi:
-            raise ValueError("mpi must be False in JAXtronomy")
+            raise ValueError("mpi must be False in JAXtronomy since parallelization is done through JAX")
 
         backend = jax.default_backend()
         if backend == "cpu":
-            num_devices = jax.device_count()
-            if n_particles % num_devices != 0:
+            if n_particles % threadCount != 0:
                 raise ValueError(
-                    f"PSO particle count {n_particles} must be divisible by the number of CPU devices for parallelization. "
-                    f"There are {num_devices} cpu devices currently recognized by JAX."
+                    f"PSO particle count {n_particles} must be divisible by threadCount for parallelization. "
+                    f"threadCount has been set to {threadCount}."
                 )
-        logL_func = prepare_logL_func(backend=backend, logL_func=self.chain.logL)
+        logL_func = prepare_logL_func(backend=backend, logL_func=self.chain.logL, threadCount=threadCount)
 
         if lower_start is None or upper_start is None:
             lower_start, upper_start = np.array(self.lower_limit), np.array(
@@ -110,7 +108,7 @@ class Sampler(Sampler_lenstronomy):
         sigma_start,
         mpi=False,
         progress=False,
-        threadCount=None,
+        threadCount=1,
         initpos=None,
         backend_filename=None,
         start_from_backend=False,
@@ -128,11 +126,11 @@ class Sampler(Sampler_lenstronomy):
         :type mean_start: numpy array of length the number of parameters
         :param sigma_start: spread of the parameter values (uncorrelated in each dimension) of the initialising sample
         :type sigma_start: numpy array of length the number of parameters
-        :param mpi: if True, initializes an MPIPool to allow for MPI execution of the sampler
+        :param mpi: if True, initializes an MPIPool to allow for MPI execution of the sampler. Must be False in JAXtronomy
         :type mpi: bool
         :param progress: if True, prints the progress bar
         :type progress: bool
-        :param threadCount: number of threats in multi-processing (not applicable for MPI)
[…]
