### 2026-04-21 b2dc72c
add error message when unsupported solver type is supplied
