### 2026-04-22 #9685 @warrickball
Hi! I've just returned from a long Easter break and will start some final editorial checks soon, which I'll probably propose via a PR to the repo. There are multiple papers in this state, so it might take a few days to get started, but I'll try my best to have this wrapped up by the end of next week.
