### 2026-04-26 9f468f8
Initial editorial tweaks for publication in JOSS
