+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @ARTICLE{Birrer:2018,
-title = {lenstronomy: Multi-purpose gravitational lens modelling software package},
+title = {{lenstronomy}: Multi-purpose gravitational lens modelling software package},
 journal = {Physics of the Dark Universe},
 volume = {22},
 pages = {189-201},
@@ -73,7 +73,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{DINOS2,
        author = {{Sheu}, William and {Shajib}, Anowar J. and {Treu}, Tommaso and {Sonnenfeld}, Alessandro and {Birrer}, Simon and {Cappellari}, Michele and {Oldham}, Lindsay J. and {Tan}, Chin Yi},
-        title = "{Project Dinos II: redshift evolution of dark and luminous matter density profiles in strong-lensing elliptical galaxies across 0.1 < z < 0.9}",
+        title = "{Project Dinos II: Redshift evolution of dark and luminous matter density profiles in strong-lensing elliptical galaxies across 0.1 < z < 0.9}",
       journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {gravitational lensing: strong, galaxies: elliptical and lenticular, cD, galaxies: evolution, cosmology: observations, dark matter, Astrophysics of Galaxies, Cosmology and Nongalactic Astrophysics},
          year = 2025,
@@ -106,7 +106,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{Ding:2020,
    author = {{Ding}, Xuheng and {Silverman}, John and {Treu}, Tommaso and {Schulze}, Andreas and {Schramm}, Malte and {Birrer}, Simon and {Park}, Daeseong and {Jahnke}, Knud and {Bennert}, Vardha N. and {Kartaltepe}, Jeyhan S. and {Koekemoer}, Anton M. and {Malkan}, Matthew A. and {Sanders}, David},
-    title = "{The Mass Relations between Supermassive Black Holes and Their Host Galaxies at 1 < z < 2 HST-WFC3}",
+    title = {The Mass Relations between Supermassive Black Holes and Their Host Galaxies at 1 < z < 2 with HST-WFC3},
   journal = {The Astrophysical Journal},
  keywords = {Galaxy evolution, Active galaxies, 594, 17, Astrophysics - Astrophysics of Galaxies},
      year = 2020,
@@ -125,7 +125,7 @@ archivePrefix = {arXiv},
 
   @ARTICLE{TDCOSMO1,
 	author = {{Millon}, Martin and {Galan}, Aymeric and {Courbin}, Frédéric and {Treu}, Tommaso and {Suyu}, Sherry H. and {Ding}, Xuheng and {Birrer}, Simon and {Chen}, G. C.-F. and {Shajib}, Anowar J. and {Sluse}, Dominique and {Wong}, Kenneth C. and {Agnello}, A. and {Auger}, M. W. and {Buckley-Geer}, Elizabeth J. and {Chan}, J. H. H. and {Collett}, T. and {Fassnacht}, Christopher D. and {Hilbert}, S. and {Koopmans}, Léon V. E. and {Motta}, Veronica and {Mukherjee}, S. and {Rusu}, Cristian E. and {Sonnenfeld}, Alessandro and {Spiniello}, Chiara and {Van de Vyvere}, Lyne},
-	title = {TDCOSMO - I. An exploration of systematic uncertainties in the inference of H0 from time-delay cosmography},
+	title = {TDCOSMO - {I}. An exploration of systematic uncertainties in the inference of H0 from time-delay cosmography},
[…]
