### 2026-04-26 #9685 @warrickball
Hi @ahuang314, I've just opened PR https://github.com/lenstronomy/JAXtronomy/pull/85 to propose some editorial tweaks.

I also noticed that @ConnorStoneAstro opened issues [#70](https://github.com/lenstronomy/JAXtronomy/issues/70), [#71](https://github.com/lenstronomy/JAXtronomy/issues/71), [#72](https://github.com/lenstronomy/JAXtronomy/issues/72) and [#73](https://github.com/lenstronomy/JAXtronomy/issues/73) in the JAXtronomy repo, I believe as part of the JOSS review. If these have been resolved (which you say in a [comment above](https://github.com/openjournals/joss-reviews/issues/9685#issuecomment-3898966670)), you can presumably close them, perhaps mentioning which commit or PR has the relevant changes.
