### 2026-04-26 lenstronomy/JAXtronomy#71 @ahuang314
Fixed in #75
