### 2026-04-26 lenstronomy/JAXtronomy#72 @ahuang314
Fixed in PR #74
