### 2026-04-26 lenstronomy/JAXtronomy#70 @ahuang314
Fixed in #75
