+++ requirements.txt
@@ -1,6 +1,6 @@
 lenstronomy @ https://github.com/lenstronomy/lenstronomy/archive/refs/heads/main.zip
 numpy>=2.0.0
-jax>=0.7.0
+jax<0.10.0
 numpyro>=0.19.0
 optax>=0.2.5
 scipy
