### 2026-04-28 a353260
add flexion to profile list base
