+++ jaxtronomy/LensModel/profile_list_base.py
@@ -13,6 +13,7 @@ _JAXXED_MODELS = [
     "EPL_MULTIPOLE_M3M4_ELL",
     "EPL_MULTIPOLE_M3M4",
     "EPL_Q_PHI",
+    "FLEXION",
     "GAUSSIAN",
     "GAUSSIAN_POTENTIAL",
     "HERNQUIST",
