+++ .pre-commit-config.yaml
@@ -25,17 +25,10 @@ repos:
     hooks:
       - id: black-jupyter
         language_version: python3
-  - repo: local
+  - repo: https://github.com/PyCQA/docformatter
+    rev: v1.7.7
     hooks:
       - id: docformatter
-        name: docformatter
-        description: Formats docstrings to follow PEP 257.
-        entry: python -m docformatter
-        additional_dependencies:
-          - docformatter==1.7.5
-          - tomli
         args: [-r, --black, --in-place]
-        language: python
-        types:
-          - python
+        language_version: python3.12
 
