### 2026-04-28 7075d67
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
