+++ jaxtronomy/LensModel/Profiles/flexion.py
@@ -38,7 +38,7 @@ class Flexion(LensProfileBase):
             * (g1 * x_**3 + 3 * g2 * x_**2 * y_ + 3 * g3 * x_ * y_**2 + g4 * y_**3)
         )
         return f_
-    
+
     @staticmethod
     @jit
     def derivatives(x, y, g1, g2, g3, g4, ra_0=0, dec_0=0):
