+++ test/test_LensModel/test_Profiles/test_flexion.py
@@ -1,4 +1,5 @@
 import jax
+
 jax.config.update("jax_enable_x64", True)
 
 import numpy as np
