### 2026-04-28 aad5246
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
